requirements-rag.txt
# RAG System Dependencies
# Install with: pip install -r requirements-rag.txt

# Vector Database
chromadb>=0.4.0

# Embedding Models
ollama>=0.1.0

# Token counting for chunking
tiktoken>=0.5.0

# YAML parsing for frontmatter
pyyaml
